ISSUE#12: Use-pen_state-in-direction-handling-second-refactor
Separated the get_pen_state function from the navigate_command_functions.
Added tests for the get_pen_state and the pen_up ignored.
Modified the test_navigate_command_functions test
Deleted the pen_up and pen_colour functions in the scope of
the test_navigate_command_functions test as they are not needed.